% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifotibble.R
\name{fifotibble}
\alias{fifotibble}
\title{Create fifo quants/valuation tibble}
\usage{
fifotibble(qty, price)
}
\arguments{
\item{qty}{vector of quantities of movements (receipts positive, issues negative)}

\item{price}{vector of prices of movements (should be same length as qty)}
}
\value{
tibble: qty, price, openstock,value (only non-zero quants with stock)
}
\description{
This function receives movements quantities and prices (purchases/sales) and returns a tibble of open stocks per remaining valuation
quants
}
\examples{
fifotibble(c(10,-5,20),c(10,12,12))
#   A tibble: 2 x 4
#qty price openstock value
#    <dbl> <dbl> <dbl> <dbl>
#1   10    10      5    50
#2   20    12     20    240
}
