% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifotibble.R
\name{gains}
\alias{gains}
\title{Create gains vector //in progress not tested}
\usage{
gains(qty, price)
}
\arguments{
\item{qty}{vector of quantities of movements (receipts positive, issues negative)}

\item{price}{vector of prices of movements (should be same lenght as qty)}
}
\value{
vector of gains
}
\description{
This function receives movements quantities and prices (purchases/selss) and returns a vector of capital gains (and NAs for qty>0)
}
