% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifotibble.R
\name{tidyfifo}
\alias{tidyfifo}
\title{Tidyverse style fifo transformation}
\usage{
tidyfifo(.data, quantities, prices)
}
\arguments{
\item{quantities}{column for quantities vector (+receipts , -issues)}

\item{prices}{column for prices vector}
}
\value{
a tibble of open quants of stock positions other columns retained as possible in; added columns: openstock, value
}
\description{
Retains only open quant rows, any extra columns retained
}
