% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifotibble.R
\name{fifotibble}
\alias{fifotibble}
\title{Create fifo quants/valuation tibble}
\usage{
fifotibble(qty, price)
}
\arguments{
\item{qty}{vector of quantities of movements (receipts positive, issues negative)}

\item{price}{vector of prices of movements (should be same length as qty)}
}
\value{
tibbles: qty_movt, price, quant_stock
}
\description{
This function receives movements quantities and prices (purchases/selss) and returns a tibble of open stocks per remaining valuation
quants
}
