% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifotibble.R
\name{gainOnSell}
\alias{gainOnSell}
\title{Calculate gain on sell (= (qty * price) - cogs )}
\usage{
gainOnSell(fifotbl, qtyToSell, price, verbose = FALSE)
}
\arguments{
\item{fifotbl}{tibble of fifo valuation quants (result of fifotibble function)}

\item{qtyToSell}{quantity of stock to sell}

\item{price}{sale price}

\item{verbose}{= F, if true returns the full tibble describing quants sold, otherwise returns just the gain/loss value}
}
\value{
tibble or single value of gain/loss
}
\description{
This function receives a quant tibble (see \code{\link{fifotibble}} ) and simulates sell of a quantity at a price, returning resulting gain or loss
}
\examples{
gainOnSell(fifotibble(c(10,-5,20),c(10,12,12)),5,12)
#[1] 10
}
\seealso{
\code{\link{fifotibble}}
}
